重新设计运行时模块
